+++ R/default-handlers.R
@@ -1,8 +1,5 @@
 #' @include plumber.R
 default404Handler <- function(req, res){
-  if (res$status == 405L) {
-    return(list(error = "405 - Method Not Allowed"))
-  }
   res$status <- 404L
   list(error="404 - Resource Not Found")
 }