+++ R/zzz.R
@@ -1,0 +1,8 @@
+
+.onLoad <- function(...) {
+
+  addDataTypeInfo_onLoad()
+
+  addParsers_onLoad()
+
+}