+++ tests/testthat/files/image.R
@@ -10,6 +10,11 @@ function() {
   plot(1:10)
 }
 
+#* @svg
+#* @get /svg
+function() {
+  plot(1:10)
+}
 
 #' @png (width = 150, height=150)
 #' @get /littlepng
@@ -20,5 +25,12 @@ function(){
 #' @jpeg (width = 150, height=150)
 #' @get /littlejpeg
 function(){
+  plot(1:10)
+}
+
+#* @svg(width=4, height=4)
+#* @get /littlesvg
+function() {
+  warning("Should not test. Image size does not decrease with dimension decrease")
   plot(1:10)
 }